+++ extensions/font-sniper/eslint.config.js
@@ -1,0 +1,4 @@
+const raycastConfig = require("@raycast/eslint-config");
+
+// Flatten any nested arrays in the config
+module.exports = raycastConfig.flat();